kafka_video/requirements.txt
kafka-python==1.4.6
opencv-python
Flask
